Split Go tools into separate directories
The previous approach of having all tools in the same go.mod had one
major problem. The tools have some common dependencies, but depend on
different *versions* of those common dependencies. Grouping them in one
go.mod forced them to use a common version of the shared dependencies,
which can change the behavior in unpredictable ways.

For example, cosign uses sigs.k8s.io/release-utils to print version
information. The version of cosign that we have installed is supposed to
be using an older version of this lib, which prints the version info to
stderr. But, by including cosign in the same go.mod file as other tools,
we got a newer version of the lib which prints to stdout. This
difference is benign, but it's easy to see how we could encounter much
more serious problems.

This also uncovered that we're installing an outdated version of cosign,
there's a 2.x available. And the same may be true for other Go tools. To
be addressed in future commits.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>